requirements.txt
pandas
numpy
gdown
Pillow
torchvision
torcheval
tensorboard
kagglehub
streamlit
opencv-python-headless
mediapipe==0.10.14
